Add basic BAI2 parser lib and tests
This is the start of a BAI2 parser library.
All the different sections are parsed and converted to python objects.

The file is parsed line by line so that if an error is found, the script
exits immediately.

There seems to be a bug in the BAI2 export we have received
but we need to investigate as it might just be my misunderstanding
of the specs.
The bug is the record 03 having an extra (empty) field.

Also, we don't have enough data to test that the different bits
work as expected so we would need to test it in pre-live and see how
it goes.

I've also refactored the code so that it's easier to reconstruct
the file back from the parsed models.